requirements.txt
numpy~=1.24.2
tqdm~=4.65.0
tensorboardX
torchvision~=0.11

pandas~=1.5.3
timm~=0.6.13